+++ wit/record.go
@@ -11,14 +11,14 @@ type Record struct {
 
 // Size returns the [ABI byte size] for [Record] r.
 //
-// [ABI byte size]: https://github.com/WebAssembly/component-model/blob/main/design/mvp/CanonicalABI.md#size
+// [ABI byte size]: https://github.com/WebAssembly/component-model/blob/main/design/mvp/CanonicalABI.md#element-size
 func (r *Record) Size() uintptr {
 	var s uintptr
 	for _, f := range r.Fields {
 		s = Align(s, f.Type.Align())
 		s += f.Type.Size()
 	}
-	return s
+	return Align(s, r.Align())
 }
 
 // Align returns the [ABI byte alignment] for [Record] r.